updated mobile_phone to BIGINT datatype, due to large numbers

+++ db/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE users(
     first_name VARCHAR(100) CHARACTER SET utf8,
     last_name VARCHAR(100) CHARACTER SET utf8,
     email VARCHAR(100) CHARACTER SET utf8 UNIQUE NOT NULL,
-    mobile_phone INT,
+    mobile_phone BIGINT,
     bed_time TIME NOT NULL,
     created_date DATETIME DEFAULT CURRENT_TIMESTAMP
 );
